fix(NanoSpec): Fix 16 C4244 function argument warnings in 4 Recipe files
- RecipeStageProgramView.cpp (line 1645)
- RecipeDistanceStageProgramView.cpp (line 1504)
- RecipeGantryStageProgramView.cpp (line 806)
- RecipeStressStageProgramView.cpp (line 2068)
- Added static_cast<int>() to Rectangle(CRect()) arguments
- Fixes C4244: 'argument': conversion from 'double' to 'int', possible loss of data

+++ SRC/NanoSpec/RecipeStressStageProgramView.cpp
@@ -2065,7 +2065,7 @@ void CRecipeStressStageProgramView::OnPaint()
 		m_dDrawEndStagePosY = StageRectBottom;
 	}
 
-	pDC->Rectangle(CRect(m_dDrawStartStagePosX, m_dDrawStartStagePosY, m_dDrawEndStagePosX, m_dDrawEndStagePosY));
+	pDC->Rectangle(CRect(static_cast<int>(m_dDrawStartStagePosX), static_cast<int>(m_dDrawStartStagePosY), static_cast<int>(m_dDrawEndStagePosX), static_cast<int>(m_dDrawEndStagePosY)));
 	pDC->SelectObject(m_oldPen);
 	pDC->SelectObject(m_oldBrush);
